# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AENS X VALURA — AI Wealth Management Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface BatchQuotesRequest {
  symbols: string[];
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type UserProfileRiskProfile =
  (typeof UserProfileRiskProfile)[keyof typeof UserProfileRiskProfile];

export const UserProfileRiskProfile = {
  conservative: "conservative",
  moderate: "moderate",
  aggressive: "aggressive",
} as const;

export type UserProfileKycStatus =
  (typeof UserProfileKycStatus)[keyof typeof UserProfileKycStatus];

export const UserProfileKycStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  riskProfile: UserProfileRiskProfile;
  kycStatus: UserProfileKycStatus;
  investmentGoal: string;
  currency: string;
  totalPortfolioValue: number;
  createdAt: string;
}

export interface Holding {
  ticker: string;
  name: string;
  shares: number;
  avgCostBasis: number;
  currentPrice: number;
  currentValue: number;
  gainLoss: number;
  gainLossPct: number;
  weight: number;
  sector: string;
  currency: string;
}

export interface Portfolio {
  userId: string;
  totalValue: number;
  cashBalance: number;
  totalCost: number;
  totalGainLoss: number;
  totalGainLossPct: number;
  holdings: Holding[];
  lastUpdated: string;
}

export type PortfolioSummaryAllocationByAssetItem = {
  name: string;
  value: number;
  pct: number;
};

export type PortfolioSummaryAllocationBySectorItem = {
  sector: string;
  pct: number;
};

export interface PortfolioSummary {
  userId: string;
  totalValue: number;
  dailyChange: number;
  dailyChangePct: number;
  totalReturn: number;
  totalReturnPct: number;
  topHolding: string;
  holdingsCount: number;
  riskScore: number;
  allocationByAsset: PortfolioSummaryAllocationByAssetItem[];
  allocationBySector: PortfolioSummaryAllocationBySectorItem[];
}

export type ConcentrationRiskFlag =
  (typeof ConcentrationRiskFlag)[keyof typeof ConcentrationRiskFlag];

export const ConcentrationRiskFlag = {
  low: "low",
  moderate: "moderate",
  high: "high",
  critical: "critical",
} as const;

export interface ConcentrationRisk {
  topPositionPct: number;
  top3PositionsPct: number;
  flag: ConcentrationRiskFlag;
}

export interface Performance {
  totalReturnPct: number;
  annualizedReturnPct: number;
  volatility: number;
}

export interface BenchmarkComparison {
  benchmark: string;
  portfolioReturnPct: number;
  benchmarkReturnPct: number;
  alphaPct: number;
}

export type ObservationSeverity =
  (typeof ObservationSeverity)[keyof typeof ObservationSeverity];

export const ObservationSeverity = {
  info: "info",
  warning: "warning",
  danger: "danger",
} as const;

export interface Observation {
  severity: ObservationSeverity;
  text: string;
}

export interface PortfolioHealthReport {
  userId: string;
  concentrationRisk: ConcentrationRisk;
  performance: Performance;
  benchmarkComparison: BenchmarkComparison;
  observations: Observation[];
  disclaimer: string;
  generatedAt: string;
}

export interface MarketQuote {
  symbol: string;
  name: string;
  price: number;
  change: number;
  changePct: number;
  volume?: number;
  marketCap?: number;
  high52w?: number;
  low52w?: number;
  currency: string;
  exchange?: string;
  lastUpdated: string;
}

export interface MarketMovers {
  gainers: MarketQuote[];
  losers: MarketQuote[];
  lastUpdated: string;
}

export interface ChatRequest {
  userId: string;
  message: string;
  conversationId?: string;
  portfolioContext?: boolean;
}

export interface Conversation {
  id: string;
  userId: string;
  title: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface Message {
  id: string;
  conversationId: string;
  role: MessageRole;
  content: string;
  intent?: string;
  agent?: string;
  createdAt: string;
}

export type ListUsers200 = {
  users: UserProfile[];
};

export type GetBatchQuotes200 = {
  quotes: MarketQuote[];
};

export type GetConversations200 = {
  conversations: Conversation[];
};

export type GetMessages200 = {
  messages: Message[];
};