# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AENS X VALURA — AI Wealth Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatRequest {
  userId: string;
  message: string;
  conversationId?: string;
  portfolioContext?: boolean;
}