# lib/api-zod/src/generated/types/marketMovers.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AENS X VALURA — AI Wealth Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MarketQuote } from "./marketQuote";

export interface MarketMovers {
  gainers: MarketQuote[];
  losers: MarketQuote[];
  lastUpdated: string;
}