# lib/api-zod/src/generated/types/observation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AENS X VALURA — AI Wealth Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ObservationSeverity } from "./observationSeverity";

export interface Observation {
  severity: ObservationSeverity;
  text: string;
}