# lib/api-zod/src/generated/types/portfolioHealthReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AENS X VALURA — AI Wealth Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BenchmarkComparison } from "./benchmarkComparison";
import type { ConcentrationRisk } from "./concentrationRisk";
import type { Observation } from "./observation";
import type { Performance } from "./performance";

export interface PortfolioHealthReport {
  userId: string;
  concentrationRisk: ConcentrationRisk;
  performance: Performance;
  benchmarkComparison: BenchmarkComparison;
  observations: Observation[];
  disclaimer: string;
  generatedAt: string;
}